feat: add zipkin image and move examples
